Added PIX performance debugging hooks (resolves issue 20). Removed auto-release of objects, and added leak tracking functionality (resolves issue 15). Renamed SlimDX.cpp (resolves issue 11).

+++ trunk/source/direct3d/GraphicsStream.cpp
@@ -151,7 +151,7 @@ namespace SlimDX
 	
 	void GraphicsStream::Write( IntPtr source, Int64 byteCount )
 	{
-		memcpy( m_Buffer + m_Position, source.ToPointer(), byteCount );
+		memcpy( m_Buffer + m_Position, source.ToPointer(), (size_t) byteCount );
 	}
 	
 	int GraphicsStream::Read( array<Byte>^ buffer, int offset, int count )
